fastapi_backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
bcrypt==4.3.0
beanie==2.0.1
certifi==2025.11.12
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.123.7
fastapi-cli==0.0.16
fastapi-cloud-cli==0.5.2
fastar==0.8.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
jinja2==3.1.6
lazy-model==0.4.0
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
motor==3.7.1
passlib==1.7.4
pycparser==2.23
pydantic==2.12.5
pydantic-core==2.41.5
pygments==2.19.2
pyjwt==2.10.1
pymongo==4.15.5
python-dotenv==1.2.1
python-multipart==0.0.20
pyyaml==6.0.3
rich==14.2.0
rich-toolkit==0.17.0
rignore==0.7.6
sentry-sdk==2.47.0
shellingham==1.5.4
starlette==0.50.0
typer==0.20.0
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
